Added comments for readability, in each file
