+++ src/main/java/entity/Pokemon.java
@@ -279,21 +279,4 @@ public class Pokemon {
     }
 
 
-    /**
-     * Prints all the statistics of the Pokemon.
-     */
-    public void printAllStats() {
-        System.out.println("Pokemon Data:");
-        System.out.println("Name: " + this.getName());
-        System.out.println("HP: " + this.getHealth());
-        System.out.println("Attack: " + this.getAttack());
-        System.out.println("Defense: " + this.getDefense());
-        System.out.println("Number: " + this.getNumber());
-        System.out.println("FrontSprite: " + this.getFrontSprite());
-        System.out.println("BackSprite: " + this.getBackSprite());
-        System.out.println("IsAlive: " + this.getIsAlive());
-    }
-
-
-
 }
