+++ src/main/java/use_case/InitializeRunGameInterface.java
@@ -4,18 +4,17 @@ import entity.PlayerorAiPokemons;
 import entity.Pokemon;
 import view.BattleView;
 import view.BattleViewInterface;
+
 import javax.swing.*;
 import java.util.List;
 
 public interface InitializeRunGameInterface {
     /**
-     * This method initializes the game state. This create player and AI object and use movesFactory to create moves.
+     * This method initializes the game state. This creates player and AI objects and uses MovesFactory to create moves.
      * Then it runs the game.
-     * */
-    static void initializeRunGame(Pokemon[] allPokemons, List<Pokemon> selectedPokemonList, JFrame frame)
-
-    {
-        // convert listPokemon to Pokemon[]
+     */
+    static void initializeRunGame(Pokemon[] allPokemons, List<Pokemon> selectedPokemonList, JFrame frame) {
+        // Convert listPokemon to Pokemon[]
         Pokemon[] selectedPokemonListArray = new Pokemon[6];
         for (int i = 0; i < 6; i++) {
             selectedPokemonListArray[i] = selectedPokemonList.get(i);
@@ -33,7 +32,7 @@ public interface InitializeRunGameInterface {
         }
 
         PlayerorAiPokemons player = new PlayerorAiPokemons(selectedPokemonListArray, "Player", 0);
-        // Random index for first Pokemon
+        // Random index for the first Pokemon
         PlayerorAiPokemons aiPlayer = new PlayerorAiPokemons(aiPokemon, "AI Player", (int) (Math.random() * 6));
 
         List<String> playerMoves = MovesFactory.createMoves(List.of(player.pokemons));
@@ -46,6 +45,5 @@ public interface InitializeRunGameInterface {
 
         BattleViewInterface battleView = new BattleView(frame);
         RunGame game = new RunGame(player, aiPlayer, battleView);
-
     }
 }
